[2~4단계 방탈출 결제 / 배포] 히스타(오승현) 미션 제출합니다.  (#310)
* feat: 클라이언트 예약 정보 관련 코드 추가

* feat: 예약 생성 시 결제 정보 저장하도록 추가

* test: 깨진 테스트 코드 수정

* feat: 결제 정보 확인 기능 구현 완

* feat: 결제 취소 이전에 예약을 삭제하지 못하도록 기능 추가

* docs: Swagger API 명세 적용

* feat: 로깅 추가

- 서비스 계층의 작업이 로깅되도록 추가
- CustomException 예외가 로깅되도록 추가

* docs: Swagger UI 각 컨트롤러에 한글 설명 추가

* test: 깨진 테스트 코드 수정

- initialize_database.sql 의 쿼리에서 completed_payment를 제일 먼저 삭제하도록 변경
- DataSourceTest.java 삭제
- 예약 정보 조회에서 결제 정보 추가

* chore: logback-spring 설정 파일 추가

- dev 환경에서는 콘솔 출력 되도록, 운영 환경에서는 파일에 기록하도록 추가

* chore: Cookie secure 설정 해제

- 현재 배포 환경(HTTP)를 고려하여 임시 해제

+++ src/main/java/roomescape/reservation/ReservationService.java
@@ -3,9 +3,11 @@ package roomescape.reservation;
 import java.time.LocalDateTime;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
+import org.hibernate.TransientObjectException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import roomescape.auth.dto.LoginMember;
+import roomescape.exception.custom.reason.reservation.ReservationNotDeletedException;
 import roomescape.payment.PaymentManager;
 import roomescape.exception.custom.reason.reservation.ReservationConflictException;
 import roomescape.exception.custom.reason.reservation.ReservationNotExistsMemberException;
@@ -15,16 +17,20 @@ import roomescape.exception.custom.reason.reservation.ReservationNotExistsTimeEx
 import roomescape.exception.custom.reason.reservation.ReservationNotFoundException;
 import roomescape.member.domain.Member;
 import roomescape.member.repository.MemberRepository;
+import roomescape.reservation.domain.CompletedPayment;
 import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.domain.ReservationDate;
 import roomescape.reservation.domain.ReservationStatus;
+import roomescape.reservation.domain.WaitingRankReservation;
 import roomescape.reservation.dto.AdminFilterReservationRequest;
 import roomescape.reservation.dto.AdminReservationRequest;
 import roomescape.reservation.dto.MineReservationResponse;
 import roomescape.reservation.dto.ReservationPaymentRequest;
 import roomescape.reservation.dto.ReservationRequest;
 import roomescape.reservation.dto.ReservationResponse;
-import roomescape.reservation.repository.ReservationRepository;
+import roomescape.reservation.repository.payment.CompletedPaymentJpaRepository;
+import roomescape.reservation.repository.payment.CompletedPaymentRepository;
+import roomescape.reservation.repository.reservation.ReservationRepository;
 import roomescape.reservationtime.domain.ReservationTime;
 import roomescape.reservationtime.repository.ReservationTimeRepository;
 import roomescape.theme.domain.Theme;
@@ -36,6 +42,7 @@ public class ReservationService {
 
     private final ReservationRepository reservationRepository;
     private final ReservationTimeRepository reservationTimeRepository;
+    private final CompletedPaymentRepository completedPaymentRepository;
     private final ThemeRepository themeRepository;
     private final MemberRepository memberRepository;
     private final PaymentManager paymentManager;
@@ -57,6 +64,10 @@ public class ReservationService {
         final Reservation savedReservation = reservationRepository.save(notSavedReservation);
 
         paymentManager.confirmPayment(request.paymentRequest());
+
+        final CompletedPayment completedPayment = CompletedPayment.of(savedReservation, request.paymentRequest());
+        completedPaymentRepository.save(completedPayment);
+
         return ReservationResponse.from(savedReservation);
     }
 
@@ -97,6 +108,7 @@ public class ReservationService {
 
     public List<MineReservationResponse> readAllMine(final LoginMember loginMember) {
         final Member member = getMemberByEmail(loginMember.email());
+
         return reservationRepository.findAllWaitingRankByMember(member).stream()
                 .map(MineReservationResponse::from)
                 .toList();
@@ -135,6 +147,10 @@ public class ReservationService {
 
         if(reservation.isPending()){
             pendingNextReservation(reservation);
+        }
+
+        if (completedPaymentRepository.existsByReservationId(id)) {
+            throw ReservationNotDeletedException.paymentNotRefunded();
         }
 
         reservationRepository.deleteById(id);
